+++ cacus/repo_daemon.py
@@ -415,6 +415,12 @@ class ApiDistroCreateHandler(ApiRequestHandler):
         simple = req['simple']
         retention = req.get('retention', 0)
         gpg_key = req.get('gpg_key', None)
+        quota = req.get('quota', None)
+
+        if quota is not None:
+            # quota management is available only for admins
+            yield self._check_token(common.Cacus.admin_access)
+
         if not simple:
             gpg_check = req['gpg_check']
             strict = req['strict']
@@ -424,7 +430,7 @@ class ApiDistroCreateHandler(ApiRequestHandler):
 
         try:
             old = yield self.settings['workers'].submit(self.settings['manager'].create_distro, distro=distro, description=description,
-                                                        components=comps, gpg_check=gpg_check, strict=strict, simple=simple,
+                                                        components=comps, gpg_check=gpg_check, strict=strict, simple=simple, quota=quota,
                                                         retention=retention, incoming_wait_timeout=incoming_wait_timeout, gpg_key=gpg_key)
             if not old:
                 self.set_status(201)
@@ -569,6 +575,25 @@ class ApiPkgRemoveHandler(ApiRequestHandler):
             self.write({'success': False, 'msg': e.message})
 
 
+class ApiPkgPurgeHandler(ApiRequestHandler):
+
+    @gen.coroutine
+    def post(self, distro=None, comp=None):
+        yield self._check_token(distro)
+        req = self._get_json_request()
+        pkg = req['pkg']
+        ver = req['ver']
+        arch = req.get('arch', None)
+
+        try:
+            r = yield self.settings['workers'].submit(self.settings['manager'].purge_package, distro=distro,
+                                                      pkg=pkg, ver=ver, arch=arch)
+            self.write({'success': True, 'msg': r})
+        except common.CacusError as e:
+            self.set_status(e.http_code)
+            self.write({'success': False, 'msg': e.message})
+
+
 class ApiPkgSearchHandler(ApiRequestHandler):
 
     @gen.coroutine
@@ -661,6 +686,7 @@ def _make_app(config):
     api_pkg_upload_re = s['repo_base'] + r"/api/v1/package/upload/(?P<distro>[-_.A-Za-z0-9]+)/(?P<comp>[-_a-z0-9]+)$"
     api_pkg_copy_re = s['repo_base'] + r"/api/v1/package/copy/(?P<distro>[-_.A-Za-z0-9]+)$"
     api_pkg_remove_re = s['repo_base'] + r"/api/v1/package/remove/(?P<distro>[-_.A-Za-z0-9]+)/(?P<comp>[-_a-z0-9]+)$"
+    api_pkg_purge_re = s['repo_base'] + r"/api/v1/package/purge/(?P<distro>[-_.A-Za-z0-9]+)$"
     api_pkg_search_re = s['repo_base'] + r"/api/v1/package/search(?:/(?P<distro>[-_.A-Za-z0-9]+))?$"
     # Distribution operations
     api_distro_create_re = s['repo_base'] + r"/api/v1/distro/create/(?P<distro>[-_.A-Za-z0-9]+)$"
@@ -679,6 +705,7 @@ def _make_app(config):
         url(api_pkg_upload_re, ApiPkgUploadHandler),
         url(api_pkg_copy_re, ApiPkgCopyHandler),
         url(api_pkg_remove_re, ApiPkgRemoveHandler),
+        url(api_pkg_purge_re, ApiPkgPurgeHandler),
         url(api_pkg_search_re, ApiPkgSearchHandler),
         url(api_distro_create_re, ApiDistroCreateHandler),
         url(api_distro_remove_re, ApiDistroRemoveHandler),