+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## [Version 0.7.17](https://github.com/beebeeep/cacus/tree/v0.7.17) (2017-09-04)
+  * Added package purging
+  * Purge old packages
+
 ## [Version 0.7.16](https://github.com/beebeeep/cacus/tree/v0.7.16) (2017-08-15)
   * fixed error handling in /api/v1/package/remove